# api_v2_backup_20251016_203324/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_EduardoMG12_cine_api_v2_internal_dto.CreateReviewRequest:
    properties:
      content:
        maxLength: 2000
        minLength: 1
        type: string
      movie_id:
        type: integer
      rating:
        maximum: 10
        minimum: 1
        type: integer
    required:
    - movie_id
    type: object
...
  github_com_EduardoMG12_cine_api_v2_internal_dto.ReviewMovieResponse:
    properties:
      adult:
        type: boolean
      backdrop_url:
        type: string
      external_id:
        type: string
      genres:
        items:
          type: string
        type: array
      id:
        type: integer
      overview:
        type: string
      poster_url:
        type: string
      release_date:
        type: string
      runtime:
        type: integer
      title:
        type: string
      vote_average:
        type: number
      vote_count:
        type: integer
    type: object
  github_com_EduardoMG12_cine_api_v2_internal_dto.ReviewResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      movie:
        $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.ReviewMovieResponse'
      rating:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.UserProfile'
    type: object
  github_com_EduardoMG12_cine_api_v2_internal_dto.UserProfile:
    properties:
      bio:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: integer
      is_private:
        type: boolean
      profile_picture_url:
        type: string
      theme:
        type: string
      username:
        type: string
    type: object
  github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse:
    properties:
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@cineverse.com
    name: CineVerse API Support
    url: http://www.cineverse.com/support
  description: A comprehensive social network platform for movie enthusiasts
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://cineverse.com/terms/
  title: CineVerse API
  version: "2.0"
paths:
...
  /reviews:
    post:
      consumes:
      - application/json
      description: Create a new review for a movie with rating and/or content
      parameters:
      - description: Review data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Review created successfully
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.ReviewResponse'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "404":
          description: Movie not found
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "409":
          description: Review already exists
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a movie review
      tags:
      - Reviews
  /reviews/{id}:
    get:
      description: Get detailed information about a specific review
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Review details
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.ReviewResponse'
        "400":
          description: Invalid review ID
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "404":
          description: Review not found
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
      summary: Get review by ID
      tags:
      - Reviews
  /reviews/movie/{movieId}:
    get:
      description: Get all reviews for a specific movie with pagination
      parameters:
      - description: Movie ID
        in: path
        name: movieId
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of movie reviews
          schema:
            items:
              $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.ReviewResponse'
            type: array
        "400":
          description: Invalid movie ID
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "404":
          description: Movie not found
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
      summary: Get movie reviews
      tags:
      - Reviews
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"